+++ src/constants/api.js
@@ -3,6 +3,7 @@ export const pageEndPoints = {
   LOADING: "/loading", // 로딩 페이지
   MYSUBSCRIPTIONS: "/feed/channels", // 모든 구독 채널 페이지
   CHANNEL: "/channel/:channel", //채널 페이지
+  TEST: "/test", // 테스트 페이지
 };
 
 export const apiEndPoints = {
