+++ src/pages/Test/Test.jsx
@@ -1,0 +1,12 @@
+import { useAuth } from "../../contexts/authContext";
+const TestPage = () => {
+  const { signIn } = useAuth();
+
+  return (
+    <div>
+      <button onClick={signIn}>로그인</button>
+    </div>
+  );
+};
+
+export default TestPage;
